+++ backend/src/main/java/com/learnbytyping/question/repository/ExamRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExamRepository extends JpaRepository<Exam, Integer> {}
+public interface ExamRepository extends JpaRepository<Exam, Long> {}
